+++ src/IODevices/PwmDevices.h
@@ -25,13 +25,13 @@ class PwmDevices : public EventListener {
     eventDispatcher->addListener(this, EVENT_SOURCE_SWITCH);
 
     // Adjust PWM properties if needed.
-    //analogWriteFreq(5000);
-    //analogWriteRange(65535);
-    //analogWriteResolution(16);
+    // analogWriteFreq(5000);
+    // analogWriteRange(65535);
+    // analogWriteResolution(16);
   }
 
-  void registerSolenoid(byte p, byte n, byte pow, byte minPT, byte maxPT,
-                        byte hP, byte hPAT, byte fS);
+  void registerSolenoid(byte p, byte n, byte pow, uint16_t minPT,
+                        uint16_t maxPT, byte hP, uint16_t hPAT, byte fS);
   void registerFlasher(byte p, byte n, byte pow);
   void registerLamp(byte p, byte n, byte pow);
 
@@ -48,10 +48,10 @@ class PwmDevices : public EventListener {
   byte port[MAX_PWM_OUTPUTS] = {0};
   byte number[MAX_PWM_OUTPUTS] = {0};
   byte power[MAX_PWM_OUTPUTS] = {0};
-  byte minPulseTime[MAX_PWM_OUTPUTS] = {0};
-  byte maxPulseTime[MAX_PWM_OUTPUTS] = {0};
+  uint16_t minPulseTime[MAX_PWM_OUTPUTS] = {0};
+  uint16_t maxPulseTime[MAX_PWM_OUTPUTS] = {0};
   byte holdPower[MAX_PWM_OUTPUTS] = {0};
-  byte holdPowerActivationTime[MAX_PWM_OUTPUTS] = {0};
+  uint16_t holdPowerActivationTime[MAX_PWM_OUTPUTS] = {0};
   byte fastSwitch[MAX_PWM_OUTPUTS] = {0};
   byte type[MAX_PWM_OUTPUTS] = {0};
   unsigned long activated[MAX_PWM_OUTPUTS] = {0};
